refactor(models): move model provider logic to models.ts
- Moved the `getModel` function from `generation.ts` to `models.ts` for better organization and separation of concerns.
- Cleaned up imports in `generation.ts` by removing unused model provider imports, enhancing clarity and maintainability.
- This change adheres to clean code principles by promoting single responsibility and improving module cohesion.
